+++ src/System.Runtime/tests/DBNullTests.netstandard.cs
@@ -15,6 +15,24 @@ namespace System.Tests
         }
 
         [Fact]
+        public static void TestToUInt64_Invalid()
+        {
+            Assert.Throws<InvalidCastException>(() => ((IConvertible)DBNull.Value).ToUInt64(null));
+        }
+
+        [Fact]
+        public static void TestToChar_Invalid()
+        {
+            Assert.Throws<InvalidCastException>(() => ((IConvertible)DBNull.Value).ToChar(null));
+        }
+
+        [Fact]
+        public static void TestToDateTime_Invalid()
+        {
+            Assert.Throws<InvalidCastException>(() => ((IConvertible)DBNull.Value).ToDateTime(null));
+        }
+
+        [Fact]
         public static void TestGetTypeCode()
         {
             Assert.Equal(TypeCode.DBNull, DBNull.Value.GetTypeCode());
@@ -25,6 +43,27 @@ namespace System.Tests
         {
             DBNull MyNull = DBNull.Value;
             Assert.True(DBNull.Value.Equals(MyNull));
+        }
+
+        [Fact]
+        public static void TestEquals_Invalid()
+        {
+            DBNull MyNull = null;
+            Assert.False(DBNull.Value.Equals(MyNull));
+        }
+
+        [Fact]
+        public static void Test_ToType()
+        {
+            Type typeParam = Type.GetType("System.String");
+            Assert.Equal(string.Empty, ((IConvertible)DBNull.Value).ToType(typeParam, null));
+        }
+
+        [Fact]
+        public static void Test_ToType_Invalid()
+        {
+            Type typeParam = Type.GetType("System.Int32");
+            Assert.Throws<InvalidCastException>(() => ((IConvertible)DBNull.Value).ToType(typeParam, null));
         }
 
         [Fact]
