+++ src/System.Runtime/tests/System/Reflection/PointerTests.cs
@@ -115,7 +115,7 @@ namespace System.Reflection.Tests
 
         [Theory]
         [MemberData(nameof(Pointers))]
-        [SkipOnTargetFramework(TargetFrameworkMonikers.UapAot, "Causes an assert to be fired at runtime in uapaot")]
+        [SkipOnTargetFramework(TargetFrameworkMonikers.UapAot, "Issue https://github.com/dotnet/corefx/issues/17450")]
         public void PointerFieldGetValue(int value)
         {
             var obj = new PointerHolder();
